moved amazon client to another directory
